+++ packages/rocketh/src/executor/types.ts
@@ -10,17 +10,24 @@ import type {
 export type DeployScriptFunction<
 	Artifacts extends UnknownArtifacts = UnknownArtifacts,
 	NamedAccounts extends UnknownNamedAccounts = UnknownNamedAccounts,
+	ArgumentsTypes = undefined,
 	Deployments extends UnknownDeployments = UnknownDeployments
-> = (env: Environment<Artifacts, NamedAccounts, Deployments>) => Promise<void | boolean>;
+> = (env: Environment<Artifacts, NamedAccounts, Deployments>, args?: ArgumentsTypes) => Promise<void | boolean>;
 
 export interface DeployScriptModule<
 	Artifacts extends UnknownArtifacts = UnknownArtifacts,
 	NamedAccounts extends UnresolvedUnknownNamedAccounts = UnresolvedUnknownNamedAccounts,
+	ArgumentsTypes = undefined,
 	Deployments extends UnknownDeployments = UnknownDeployments
 > {
-	(env: Environment<Artifacts, ResolvedNamedAccounts<NamedAccounts>, Deployments>): Promise<void | boolean>;
+	(env: Environment<Artifacts, ResolvedNamedAccounts<NamedAccounts>, Deployments>, args?: ArgumentsTypes): Promise<
+		void | boolean
+	>;
 	providedContext: ProvidedContext<Artifacts, NamedAccounts>;
-	skip?: (env: Environment<Artifacts, ResolvedNamedAccounts<NamedAccounts>, Deployments>) => Promise<boolean>;
+	skip?: (
+		env: Environment<Artifacts, ResolvedNamedAccounts<NamedAccounts>, Deployments>,
+		args?: ArgumentsTypes
+	) => Promise<boolean>;
 	tags?: string[];
 	dependencies?: string[];
 	runAtTheEnd?: boolean;