+++ packages/rocketh-deploy/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-deploy
 
+## 0.6.5
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.6.3
+
 ## 0.6.4
 
 ### Patch Changes